Started implementing the calculations for advancing the vehicle down the trail, it will trigger event to attach new game mode or alter travel mode state when the next location is reached. Progress towards the next location is shown as miles remaining and vehicle ticks up total distance traveled on property odometer. I also started implementing some of the formulas from the old versions of the game that I reverse engineered and it's working well!

+++ Src/TrailEntities/Trail/ForkInRoad.cs
@@ -9,7 +9,7 @@ namespace TrailEntities
         /// <summary>
         ///     Initializes a new instance of the <see cref="T:TrailEntities.ForkInRoad" /> class.
         /// </summary>
-        public ForkInRoad(string name, ulong distanceLength, IEnumerable<Location> skipChoices)
+        public ForkInRoad(string name, int distanceLength, IEnumerable<Location> skipChoices)
             : base(name, distanceLength)
         {
             _skipChoices = new HashSet<Location>(skipChoices);
